update confirm delete user

+++ usermanagement.php
@@ -1,6 +1,6 @@
 <?php
 $pagename = "User Management"; // INI "Profile" CONTOH DOANK, NANTI KALIAN GANTI SENDIRI DENGAN NAMA PAGE YANG KALIAN BUAT 
-$urlname = "index.php"; // INI "index.php" CONTOH DOANK, NANTI KALIAN GANTI SENDIRI DENGAN URL PAGE YANG KALIAN BUAT 
+$urlname = "usermanagement.php"; // INI "index.php" CONTOH DOANK, NANTI KALIAN GANTI SENDIRI DENGAN URL PAGE YANG KALIAN BUAT 
 require 'database/config.php'; // config buat koneksi database doank
 require 'authentication.php'; // authentication buat atur session, dll
 
@@ -48,7 +48,7 @@ require 'features/sidebar.php';
                 <?php foreach ($tabelevent as $row) : ?>
                     <tr>
                         </td>
-                        <td><a href="deleteuser.php?user_id=<?= $row["user_id"]; ?>"" class=" btn btn-danger btn-sm">Delete User account</a></td>
+                        <td><a href="deleteuser.php?user_id=<?= $row["user_id"]; ?>" onclick="return confirm('Are you sure want to delete this account?');" class=" btn btn-danger btn-sm">Delete User account</a></td>
                         <td><?= $row["user_name"]; ?></td>
                         <td><?= $row["user_email"]; ?></td>
                         <td><?= $row["user_password"]; ?></td>
